requirements.txt
torch==2.2.0
opencv-python
torchvision==0.17
tqdm
numpy
PyQt5
PyQt5-tools
pyserial